two major changes: - GenBranches now runs over HTo2XTo2Mu2J as well as HTo2XTo4Mu - GenOnlyNTupler has been removed completely; SimpleNTupler takes parameters controlling which branches to write
Detailed Notes:

GenBranches::Fill now takes an additional argument, "finalState", which takes effect if isSignal is true, and is either "4Mu" or "2Mu2J". This determines which sub-function is called; the gen_* branches are filled differently depending on what the input is (2 kinds of signal, or background MC). GenBranches.h and GenBranches.cpp reflect this change. Additionally, the sub-functions are defined in GenBranches.cpp, including the new one for HTo2XTo2Mu2J.

SimpleNTupler now takes two additional parameters from the cmsRun config file: "finalState" and "source". finalState is the string mentioned above, and is passed directly to GenBranches::Fill. source is one of PAT, AOD, or GEN, and completely controls which branches get declared and filled. since the EDAnalyzer doesn't complain if the input collection doesn't exist, all that is needed is to make sure that the relevant getByToken and Fill functions are not called if the input source doesn't have the collections. so GEN will only run the GEN part, AOD will run everything except the PAT part, and PAT will run everything. now, there is no need for the GenOnlyNTupler, which was always the GEN subset of SimpleNTupler. The SimpleNTupler_cfi initializer adds these parameters and removes the old commented lines, which are no longer necessary.

The NTupler_cfg no longer needs the PLUGINS parameter, but the FINALSTATE, SOURCE, and GENS_TAG parameters have been added. As before, everything gets set automatically by runNTupler. Everywhere there was PLUGINS before, now just says SimpleNTupler, since that is now the only NTupler. runNTupler has been changed to deal with the new parameters and remove the old.

CFGParser was modified to handle all the new parameters, and to automatically set the input dataset based on the value of SOURCE. This gets set automatically with either nothing, --genonly, or --aodonly (a new option), and the file declares which dataset (part of the dataHandler) to use. this streamlines much of the rest of the code, since guesswork is no longer necessary to know which process or dataset or file name to use: it's explicit.

dataHandler.py added an extra line to error output.
BuildFile.xml removed the GenOnlyNTupler.
READMEs have been changed to reflect all this.