update to Analysis with test scripts and updates for newer tree format
- Primitives was updated for newer trees, including d0 dz, sig, bs info, etc.
- testPrimitives is a new testing script for validating the Primitives library
- dataHandler adds a short phrase to an error message
- CFGParser handles the dataset dict more gracefully
- testTupler is a new testing script for validating the Tupler ecosystem.
Except for signal PAT, it should run without errors and produce several
small ROOT files with tuples, for each of the genonly, aodonly and pat modes.
- all analyzers in Analysis/analyzers updated for LXY() vs Lxy, d0 access,
and GEN getPrimitives.
- all analyzers have a FILE argument setting it to aodOnly_ blah blah, this
is temporary until I have nTuples that are made from PAT Tuples. needs to be
changed manually for now, maybe I can add an argument to Analyzer. or DH
- nMinusOne adds d0 now, since Slava has written the correct code for it
- genPlots changes the eta axes to -5 to 5 instead of 0 to 5
- Analyzer now had 3 choices of F_NTUPLE for 3 different prefixes. choose
the correct one from the relevant analyzer constructor, for now. or DH.
- Selections also updated for LXY() and d0Sig()